chore: Move types to separate folder
