+++ app/javascript/src/types/wiki.d.ts
